requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# hee-case-study-migration/src/case_study_brxm_yaml_decorator.py: 4
Unidecode == 1.2.0

# hee-case-study-migration/src/case_study_brxm_yaml_decorator.py: 8
requests == 2.25.1

# hee-case-study-migration/src/main.py: 1
ruamel.yaml == 0.17.7

# hee-case-study-migration/src/case_study_extractor.py: 1
sortedcontainers == 2.4.0

# hee-case-study-migration/src/case_study_brxm_yaml_decorator.py: 3,9
tika == 1.24
